+++ packages/andrjs/CHANGELOG.md
@@ -1,12 +1,22 @@
 # @andromedaprotocol/andromeda.js
 
+## 2.0.0-beta.3
+
+### Patch Changes
+
+- NPM package updates
+- Fix: Denom utilities now return the correct amount of decimal places
+- Tests: Added tests for denom utilities
+- Fix: Query Contract Raw supports Uint8Array keys now
+- Updated dependencies
+  - @andromedaprotocol/gql@1.0.1-beta.1
+
 ## 2.0.0-beta.2
 
 ### Patch Changes
 
 Feat - Endpoint can now be rpc string or RPC Client
 Feat - Added simulateRaw command
-
 
 ## 2.0.0-beta.1