+++ packages/cli/CHANGELOG.md
@@ -1,9 +1,20 @@
 # @andromedaprotocol/cli
 
-## 2.0.0-beta.2
+## 2.0.0-beta.3
 
 ### Patch Changes
 
+- NPM package updates
+- Wallet Address cache
+- GQL Assets default to app-contract
+- wasm query-raw command changed to query raw key
+- wasm contract-state command now list all contract keys (previously knows as wasm query-raw command)
+- Updated dependencies
+  - @andromedaprotocol/andromeda.js@2.0.0-beta.3
+
+## 2.0.0-beta.2
+
+### Patch Changes
 
 Feat - wasm query-raw and info commands
 Feat - Legacy wallet can be imported in any environment